+++ backend/build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.6'
 }
 
-group = 'com.ohgiraffers'
+group = 'com.beauty4u'
 version = '0.0.1-SNAPSHOT'
 
 java {
