+++ backend/src/main/java/com/beauty4u/backend/BackendApplication.java
@@ -1,4 +1,4 @@
-package com.ohgiraffers.backend;
+package com.beauty4u.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
