+++ backend/src/main/java/com/beauty4u/backend/common/common.java
@@ -1,4 +1,4 @@
-package com.ohgiraffers.backend.common;
+package com.beauty4u.backend.common;
 
 // common 폴더 테스트용 파일
 public class common {
