+++ backend/src/main/java/com/beauty4u/backend/config/config.java
@@ -1,4 +1,4 @@
-package com.ohgiraffers.backend.config;
+package com.beauty4u.backend.config;
 
 // config 폴더 테스트 파일
 public class config {
