+++ backend/src/main/java/com/beauty4u/backend/test/application/controller/TestController.java
@@ -1,0 +1,4 @@
+package com.beauty4u.backend.test.application.controller;
+
+public class TestController {
+}
