+++ backend/src/main/java/com/beauty4u/backend/test/application/dto/TestDTO.java
@@ -1,0 +1,4 @@
+package com.beauty4u.backend.test.application.dto;
+
+public class TestDTO {
+}
