+++ backend/src/main/java/com/beauty4u/backend/test/application/service/TestService.java
@@ -1,0 +1,4 @@
+package com.beauty4u.backend.test.application.service;
+
+public class TestService {
+}
