+++ backend/src/main/java/com/beauty4u/backend/test/domain/aggregate/Test.java
@@ -1,0 +1,4 @@
+package com.beauty4u.backend.test.domain.aggregate;
+
+public class Test {
+}
