+++ backend/src/main/java/com/beauty4u/backend/test/domain/repository/TestRepository.java
@@ -1,0 +1,4 @@
+package com.beauty4u.backend.test.domain.repository;
+
+public interface TestRepository {
+}
