+++ backend/src/main/java/com/beauty4u/backend/test/domain/service/TestDomainService.java
@@ -1,0 +1,4 @@
+package com.beauty4u.backend.test.domain.service;
+
+public class TestDomainService {
+}
