--- backend/src/main/java/com/ohgiraffers/backend/test/application/controller/TestController.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.backend.test.application.controller;
-
-public class TestController {
-}
