--- backend/src/main/java/com/ohgiraffers/backend/test/application/dto/TestDTO.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.backend.test.application.dto;
-
-public class TestDTO {
-}
