--- backend/src/main/java/com/ohgiraffers/backend/test/application/service/TestService.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.backend.test.application.service;
-
-public class TestService {
-}
