--- backend/src/main/java/com/ohgiraffers/backend/test/domain/aggregate/Test.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.backend.test.domain.aggregate;
-
-public class Test {
-}
