--- backend/src/main/java/com/ohgiraffers/backend/test/domain/repository/TestRepository.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.backend.test.domain.repository;
-
-public interface TestRepository {
-}
