--- backend/src/main/java/com/ohgiraffers/backend/test/domain/service/TestDomainService.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.backend.test.domain.service;
-
-public class TestDomainService {
-}
