+++ backend/src/test/java/com/beauty4u/backend/BackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ohgiraffers.backend;
+package com.beauty4u.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
